Move dependency on pyo3 in pure rust file to where it's actually consumed
